app: aboot: modify api declaration
In check_aboot_addr_range_overlap the first argument
should be changed to uintptr_t type instead of uint32_t.
Becasue its been called from multiple places with void pointer
type casted to uint32_t, which is invalid conversion

(cherry picked from commit bf6dfda2281f6e8513bbbe43d20e9d30ab1ce410)

+++ app/aboot/aboot.c
@@ -671,7 +671,7 @@ unsigned page_mask = 0;
  * start: Start of the memory region
  * size: Size of the memory region
  */
-int check_aboot_addr_range_overlap(uint32_t start, uint32_t size)
+int check_aboot_addr_range_overlap(uintptr_t start, uint32_t size)
 {
 	/* Check for boundary conditions. */
 	if ((UINT_MAX - start) < size)
@@ -926,7 +926,7 @@ int boot_linux_from_mmc(void)
 
 		dprintf(INFO, "Loading boot image (%d): start\n", imagesize_actual);
 
-		if (check_aboot_addr_range_overlap((uint32_t)image_addr, imagesize_actual))
+		if (check_aboot_addr_range_overlap((uintptr_t) image_addr, imagesize_actual))
 		{
 			dprintf(CRITICAL, "Boot image buffer address overlaps with aboot addresses.\n");
 			return -1;
@@ -943,7 +943,7 @@ int boot_linux_from_mmc(void)
 
 		offset = imagesize_actual;
 
-		if (check_aboot_addr_range_overlap((uint32_t)(image_addr + offset), page_size))
+		if (check_aboot_addr_range_overlap((uintptr_t)image_addr + offset, page_size))
 		{
 			dprintf(CRITICAL, "Signature read buffer address overlaps with aboot addresses.\n");
 			return -1;
